refactor(tabs): Remove wiki steps from calculator flow
- BudgetTab: Render BudgetInput directly (no wiki step)
- StackTab: Remove all wiki screens, navigate between inputs only
- DealTab: Render DealInput directly (no wiki step)

Calculator now pure action zone - education lives in /intro wiki
